refactor: update application mutation
